UPDATE: Add new file to incorporate jQuery and an id to target a specific element using jQuery
Add an id to the span element so that it can be selected by its id to
use the jQuery selector to change its inner text.

Add a new file jQuery.js to incorporate jQuery in our lms code. This is done
to establish a connection between a click event on DOM and the lms footer.
Clicking on the button adds a dynamic class to the footer which further changes
styling to make it responsive.
JQuery code also targets a span element on dom using its id and changes its innerText.

On branch uiChanges
Changes to be committed:
	modified:   course/templates/completion_automatic.mustache
	new file:   theme/moove/javascript/JQuery.js